pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer"
authors = ["krushovice <kickstar69@yandex.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.8.1"
Flask = "2.2.2"
gunicorn = "20.1.0"
python-dotenv = "0.21.0"
poetry = "1.6.1"
psycopg2-binary = "2.9.7"
pytest-cov = "^4.1.0"
beautifulsoup4 = "4.12.2"
Werkzeug = "2.2.2"
validators = "^0.22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[tool.poetry.scripts]
app = "page_analyzer.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
